requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
psycopg2 
uvicorn
python-dotenv